Split text generation module into separate files for better maintainability and organization
